+++ scripts/lichessTools-init.js
@@ -132,6 +132,7 @@ lichessTools.loadTool(LiChessTools.Tools.BetterBestArrow);
 lichessTools.loadTool(LiChessTools.Tools.DownloadBroadcastPgn);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerMoreGames);
 lichessTools.loadTool(LiChessTools.Tools.CustomBoardImage);
+lichessTools.loadTool(LiChessTools.Tools.StudyTopics);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);