+++ HISTORY.md
@@ -3,6 +3,7 @@ History of features added to LiChess Tools in time (reversed order)
 13 Oct
 
 - bulk edit chapter PGN tags
+- sortable study topics
 
 11 Oct